# specs/002-web-todo-app/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Application API
  version: 2.0.0
  description: |
    Phase II Web Application API - RESTful endpoints for task management with authentication.

    **Authentication**: JWT tokens stored in HTTP-only cookies
    **Base URL**: https://api.example.com (to be replaced with actual deployment URL)

  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server (to be updated)

tags:
  - name: Authentication
    description: User registration, login, and logout
  - name: Tasks
    description: Task CRUD operations

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with email and password
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              validRegistration:
                summary: Valid registration
                value:
                  email: user@example.com
                  password: SecurePass123
      responses:
        '201':
          description: User created successfully, JWT token set in cookie
          headers:
            Set-Cookie:
              description: JWT token in HTTP-only cookie
              schema:
                type: string
                example: token=eyJhbGc...; HttpOnly; Secure; SameSite=Strict; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailExists:
                  summary: Email already registered
                  value:
                    detail: Email already registered
                    field: email
                invalidEmail:
                  summary: Invalid email format
                  value:
                    detail: Invalid email format
                    field: email
                weakPassword:
                  summary: Password too short
                  value:
                    detail: Password must be at least 8 characters
                    field: password

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate with email and password, receive JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, JWT token set in cookie
          headers:
            Set-Cookie:
              description: JWT token in HTTP-only cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Invalid email or password

  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Clear JWT token cookie and end session
      operationId: logoutUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful, cookie cleared
          headers:
            Set-Cookie:
              description: Expired cookie to clear token
              schema:
                type: string
                example: token=; HttpOnly; Secure; SameSite=Strict; Max-Age=0
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully

  /api/tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks for authenticated user
      description: Retrieve paginated list of tasks owned by the current user
      operationId: listTasks
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of tasks per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: List of tasks with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Tasks
      summary: Create new task
      description: Create a new task for the authenticated user
      operationId: createTask
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              withDescription:
                summary: Task with description
                value:
                  title: Buy groceries
                  description: Milk, eggs, bread
              titleOnly:
                summary: Task without description
                value:
                  title: Call dentist
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyTitle:
                  summary: Empty title
                  value:
                    detail: Title cannot be empty
                    field: title
                titleTooLong:
                  summary: Title exceeds 200 characters
                  value:
                    detail: Title must be between 1 and 200 characters
                    field: title
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get single task
      description: Retrieve details of a specific task owned by authenticated user
      operationId: getTask
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Task not found

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update title and/or description of a task
      operationId: updateTask
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Permanently delete a task
      operationId: deleteTask
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tasks/{id}/complete:
    patch:
      tags:
        - Tasks
      summary: Toggle task completion
      description: Mark task as complete or incomplete (toggles current state)
      operationId: toggleTaskCompletion
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      tags:
        - System
      summary: Health check endpoint
      description: Check if the API is running
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HTTP-only cookie

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: SecurePass123
          description: Password must be at least 8 characters

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: SecurePass123

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        created_at:
          type: string
          format: date-time
          example: 2026-01-12T10:30:00Z

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          default: ""
          example: Milk, eggs, bread

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Buy groceries
          nullable: true
        description:
          type: string
          maxLength: 1000
          example: Milk, eggs, bread, cheese
          nullable: true
      description: At least one field must be provided

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          example: Milk, eggs, bread
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: 2026-01-12T10:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-01-12T10:35:00Z

    TaskListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 237
          description: Total number of tasks for the user
        pages:
          type: integer
          example: 5
          description: Total number of pages

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Error message describing what went wrong
        field:
          type: string
          nullable: true
          example: email
          description: Field name if validation error